Update html.test.ts
The test that was failing in Travis CI for VSCode was expecting 10 but getting 33. However, I'm not sure if this is a simple assertion error (in which case this should fix it) or if it is an actual failure in the feature that is correctly picked up by the tests (in which case please reject this PR).